Backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
requests>=2.31.0
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0
pydantic-settings>=2.2.1
SQLAlchemy>=2.0.0

